campaign_brain/requirements.txt
# Core dependencies
langgraph==0.2.45
langchain==0.3.4
langchain-openai==0.2.2
openai==1.51.2

# Data handling
pandas==2.2.3
numpy==1.26.4

# Environment and configuration
python-dotenv==1.0.1

# Web scraping and APIs
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.0

# Analytics and visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# UI (Streamlit for dashboard)
streamlit==1.39.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Development tools
black==24.10.0
flake8==7.1.1

# Additional utilities
pydantic==2.9.2
typing-extensions==4.12.2